refactor: 상수 파일 별도 폴더로 추출
- Constants.js 파일을 별도의 constants 폴더로 추출하여 사용
